refactor(ops): add doc and make some changes
