+++ src/test/scala/scatan/model/game/BaseScatanStateTest.scala
@@ -5,7 +5,7 @@ import scatan.model.components.BuildingType
 import scatan.model.game.config.ScatanPlayer
 import scatan.model.game.state.ScatanState
 import scatan.model.game.state.ops.BuildingOps.assignBuilding
-import scatan.model.game.state.ops.EmptySpotsOps.{emptyRoadSpots, emptySpots}
+import scatan.model.game.state.ops.EmptySpotOps.{emptyRoadSpots, emptySpots}
 import scatan.model.map.{RoadSpot, Spot, StructureSpot}
 
 abstract class BaseScatanStateTest extends BaseTest:
